test: 테스트 케이스 작성 - 구매한 각 로또 번호가 1~45 사이의 중복되지 않은 숫자인지 확인한다.
