+++ src/js/__tests__/lotto.test.js
@@ -1,48 +1,3 @@
-import {
-  isPositiveInteger,
-  isDivisibleBy,
-  createRandomNumber,
-  createRandomNumberList,
-} from '../app.js';
-
-test('구입 금액은 양의 정수이다. 실패 케이스', () => {
-  const payment = -1;
-
-  expect(() => {
-    isPositiveInteger(payment);
-  }).toThrowError();
-});
-
-test('구입 금액은 양의 정수이다. 실패 케이스', () => {
-  const payment = ' ';
-
-  expect(() => {
-    isPositiveInteger(payment);
-  }).toThrowError();
-});
-
-test('구입 금액은 양의 정수이다. 성공 케이스', () => {
-  const payment = 3;
-
-  expect(isPositiveInteger(payment)).toBe(true);
-});
-
-test('구입 금액이 1000으로 나누어 떨어진다. 실패 케이스', () => {
-  const payment = 33;
-  const lottoPrice = 1000;
-
-  expect(() => {
-    isDivisibleBy(payment, lottoPrice);
-  }).toThrowError();
-});
-
-test('구입 금액이 1000으로 나누어 떨어진다. 성공 케이스', () => {
-  const payment = 3000;
-  const lottoPrice = 1000;
-
-  expect(isDivisibleBy(payment, lottoPrice)).toBe(3);
-});
-
 expect.extend({
   toBeWithinRange(received, floor, ceiling) {
     const pass = received >= floor && received <= ceiling;
@@ -62,14 +17,13 @@ expect.extend({
   },
 });
 
-test('구입한 로또의 각각의 번호가 1~45 사이의 숫자이도록 한다. 성공 케이스', () => {
-  const randomNumber = createRandomNumber(1, 45);
+test('구매한 각 로또 번호가 1 ~ 45 사이의 중복되지 않은 숫자인지 확인.', () => {
+  const lotto = new Lotto();
 
-  expect(randomNumber).toBeWithinRange(1, 45);
-});
+  lotto.setLotto();
+  const lottoList = lotto.getLotto();
 
-test('구입한 로또 번호는 서로 다른 랜덤한 숫자 6개로 이루어진 값이다, 성공 케이스', () => {
-  const randomNumberList = createRandomNumberList(6);
-
-  expect(randomNumberList.length).toBe(new Set(randomNumberList).size);
+  lottoList.forEach((lottoNumber) => {
+    expect(lottoNumber).toBeWithinRange(1, 45);
+  });
 });
